Putting tests for mobile header in another file
